+++ NEWS.md
@@ -15,6 +15,7 @@ Version 1.0.0rc1 - dev
   and also adds support for AppImages provided by ODA.
 - CHANGE: removed deprecated features
 - CHANGE: type annotation refactoring
+- CHANGE: Table drawing add-on, R12 MText add-on
 - BUGFIX: [#747](https://github.com/mozman/ezdxf/issues/747)
   fix virtual entities of 3D DIMENSION entities  
 - BUGFIX: [#748](https://github.com/mozman/ezdxf/issues/748)